Altered ids.xml
When I attempted to run my GoalCreation part it gave me an error saying that something was wrong with the ids.xml file. So I altered it so that each item does not have text context. When looking this up it appears that AndroidStudio likes the format without text context more.