Goal Creation M2 done
My section "Goal Creation" runs when values are put into all the fields. Gives notifications if a requirement is not met (need to give goal a name). And for the current moment it will also let the user know, through a notification, that a new goal was made since actively saving and giving that info to the main list area isn't possible yet.